initial mineral gkm implementation
